JDBC: properly advertise Connection read-only mode
JDBC's Connection read-only mode appears to be just a hint to the driver
to enable certain optimisations, if possible.  It has nothing to do with
being able to write to the database (or make modifications).  Hence, a
Connection with MonetDB is never read-only, and setReadOnly should only
emit a warning when called with true.

Thanks to Martin Weindel for pointing this out in bug #2818.