configure: clean up GCC (warning) flag selection for --enable-strict
Introduced a macro to check if the compiler groks a given flag, such
that we don't have to maintain a large knowledge base of what compilers
where are known/expected to support something.  Instead we just test if
it works (compile time).

Cleaned up a lot of legacy that we don't/can't test, or is too old to
consider supporting.