mclient: report URI to database when available
Instead of just reporting the database name, report the full URI to the
database when available (under merovingian).  Since mclient "searches"
for local merovingians, and merovingians also connect to "foreign"
databases, a simple mclient connect may end up nearly anywhere.  In such
cases it is useful to have a full URI of the database that one
eventually got connected to.