Removed conjunction which is always true.
Programming 101.
Put the right-hand side of #define statements in parentheses if they
are to be used as expressions.

Consider the define:
#define isScalar(C)  C->adt != TYPE_bat
and consider how it is used:
if (!isScalar(c) && c->c[0])

This expands to
if (!c->adt != TYPE_bat && c->c[0])
!c->adt is either 0 or 1 (by definition of the ! operator), and
TYPE_bat is equal to 5.  Both 0 and 1 are not equal to 5, so the
first part of the conjunction evaluates to 1 (true).

(The obvious fix, putting the right-hand side of the isScalar define
in parentheses, changed the outcome of the if test which resulted in
failing tests.)