MonetConnection: serialise access to the connection to the server
It seems at some point in time we destroyed the locks around the
"server" connection object that was meant to make sure only one thread
at a time would do a question-response sequence.  Since the lock around
the code in the main loop was missing, race-conditions could take
please, which lead to unexplainable and odd errors in case of
multithreaded use of a single connection.