parsing: rewrite scanner and parser to explicitly ask for JSON fragments
We need to have access to [ and { chars in the parser, for things like
"var[x]", "sort by [<expr>]" or "transform {$t}" syntax.  Make it
explicit when the parser expects a JSON bit (because parsing of that is
deferred to the JSON shredder) so we can prevent parsing JSON twice.