transform: optimise variable referencing space-wise
There is no need to duplicate the values (and their kind reference) when
we just reference a variable.  Since one of the assumptions is that all
BATs are immutable, we can reference any value as many times as we want
without problems.
This just reduces code size, but also reduces copying work and hence
data size.