expand: pull out "inner pipes"
expand knows an odd syntax where it runs for each of its input arrays
first an "inner pipe" on the array, before its element members are
promoted to the top level.  While this "feature" seems to have no
benefit at all in functionality (it is as easily written as a top-level
pipe following the expand), it probably exists to steer code generation
for Hadoop's map-reduce jobs, like top and topn also do.  Since we don't
do that, we just rewrite "inner pipes" to become top level ones, and
leave optimisation to the engine itself, instead of having the user to
explicitly perform/know it.