test scripts: don't force possibly non-existing locale
en_US.UTF-8 doesn't always exist (mostly because I forgot to install it
on zonnestraal), so don't force using it.  Instead, we should make sure
that all of our build hosts set a proper UTF-8 locale themselves (in
fact, most already do).

We might want to change Mtest to check for this, and possibly try to
find a suitable UTF-8 locale.