mclient: avoid misalignment when tabs are in the string values
When a tab occurs in the tuple values, misalignment would occur when
printing the tabular grid, since the tab takes only one "character" in
width calculations, but may result in at max 8 when printing to the
screen.
Taking into account that a tab can be 1 to 8 characters when printing,
based on the starting position is hard, so alternative would be to
replace the tab with multiple (a fixed amount of) spaces instead.  Doing
this, however, makes calculating the length of the strings, and places
to cut off (also taking into account UTF-8 encoding of the strings) very
complex.  It advocates to do a global replacement of all tabs with the
desired amount of spaces a priori, which requires expensive (re)mallocs.

Hence, instead, we just replace all tab-characters with a single space,
such that the originally reported width (by the server), as well as the
calculated utf8strlen (which both count that tab as one "character")
match with what we print, and hence all alignment calculations can
remain unmodified and result in the correct offsets.