Iterator semantics changes The iterators over BATs and scalar ranges have been changed as part of the de-Mx activities. Iterators at the MAL level are concentrated in the new iterator module. They can be used as follows to construct for i:=0 step 1 until 10:
barrier i:=0;
   redo i:= iterator.next(1,10);
exit;

and for BAT variables:

barried (h,t) := iterator.new(b);
  redo (h,t) := iterator.next(b);
exit

MAL interpreter semantics to deal with barrier statements has
been changed. Now any range can be constructed, as the barrier
is left with the control variable (the first in the target list)
is set to nil. For boolean variables, the barrier is only
entered when it is set to true.

The dataflow() returns a bit.

Most changes here reflect the carry thru of this change into
the test set. (a few requires some more testing)