Prevent a race condition from happening by copying trace bats.
In e.g. the mat.slice_limit1.Bug-2645 test, trace information is
retrieved and then used in SQL queries.  This could lead to a race
condition, since the trace information was returned as references to
the actual bats which were then used in e.g. a join.  The join
happened to create a temporary view on the bat which caused the share
count of the trace bat to be 1 temporarily.  In the mean time, in
another thread, MAL instructions were being executed.  Whenever an
instruction executes, an entry is added to the trace bat.  This means
writing to the bat, which is only allowed if, among other tests, the
share count is 0.  If the bat append happens just as the view exists,
the append fails with an error.

The solution is to not return references to the trace bats, but to
return copies instead.