parse_json_string: make sure we propagate any changes made by stream realloc
parse_json_pair needs to do a string parse, followed by another value
parse.  Since that string ends up in the name BAT with the oid of
another value, and not in the string BAT with an oid of itself,
parse_json_pair needs to make sure that parse_json_string does not
change the p pointer (or its saved x variant), or the string reference
will be wrong.  For that parse_json_string needs to know this pointer,
and change it accordingly.