sabaoth: change what information is serialised
This changeset introduces changes necessary for the monetdb utility to
show a connection uri for each database with the status command.
Because this required a change in how the sabaoth data is serialised,
the protocol version has been increased, and some pending changes
applied at the same time.

- new state SABdbStarting indicates a database is starting up, but not
  yet started
- removal of path and conns properties in serialisation since they are
  not necessary and even compromising at a remote site
- introduction of dbname and uri properties to identify the database

msab_deserialise() has support for parsing both protocol 1 and protocol
2 strings, turning them into the new sabaoth structure, with necessary
omission of data for protocol 1.  This allows monetdb still to be used
with older monetdbd.