PreparedStatement: don't mix up parameters and columns
PreparedStatements can return metadata for parameters as well as result
columns since Dec2011.  In previous fixes to cope with this, I didn't
quite realise that there is a difference between the two.  We now
properly separate both, such that we don't mistakenly claim a column or
parameter doesn't exist, such as in bug #3192.  This changeset makes
Test_PSmetadata working correctly.