Concurrency on namespace
This patch addresses the bug #3163
The concurrency conflict has been addressed in the namespace in Oct branch.

Running on my desktop and a small version of the database (5000),
which is enough the create the load and 100 concurrent users
(of which 64 are accepted) does not crash the server.

However, if you run the script with a naively large sequence (eg. 1000)
you will encounter bash/OS fork/resource  limitations.