Mtest: use UNIX socket when the platform is capable
Try to avoid our socket-spam a bit for the faster platforms we have in
use.  On those we see socket bind problems, and a very large amount of
TIME_WAIT connections.  By using a UNIX socket connection we half this,
since only the server sets up the connection, but the client no longer
uses it.  Reason to still setup the socket is for python/java/etc
applications.

Refinement would be to actually not setup the TCP socket when we know
we're running an mclient test (we do know that), however that requires
some more efforts.  Let's first see if this does the trick already.