replace static by (IMHO much more informative) dynamic colormap:
Instead of an arbitrarily chosed stay map of functions to colors,
we now start witn an empty map of as many slots as we have colors
(currently 141), and collect functions as they come, until the
map is full, aggregating all remaining functions in "*.*".

Note that all of TPCH (incl. load and keys) calls only
101 different MAL functions in total.

We then show only the most expensive (timeused) functions
in the legend, ordered by desceding timeused;
if there are more than fit with freq > 0,
we show the remaining ones aggregated as "*.*".