added test for bug 3277:
The problem is threefold:

1)
As of the Feb2013 release of MonetDB,
stddev() has been replaced by stddev_pop() & stddev_samp(),
i.e., there is no stddev() any more.
We might have failed to communicate this change clearly.
Sorry!

2)
stddev_pop() & stddev_samp() fail on the results of count(),
as the latter returns MonetDB internal type wrd, which is
not a standard/valid SQL type, and (thus?) there is not
implementation of stddev_pop() & stddev_samp() for type wrd.
Explicitly casting the result of count() to a valid standard
SQL integer type (tinying, smallint, integer, bigint) makes
these functions work fine.
Open question is, whether we should simple "treat the symptom"
by adding implementations for stddev_pop() & stddev_samp() on
type wrd, or rather cure the cause by making count() return a
valid standard SQL type, e.g., integer on 32-bit systems and
bigint on 64-bit systems.

3)
sign() seems to fail as the created physical plan tries to
insert a value of type bte into a BAT with tail type int,
and while this should be no problem, there seems to be
no signature that allows this.
We'll need to investigate whether simply adding that
signature is a valid solution, or whether there is a need
to fix the translation of sign().