BBPtrim triggered relative to resource usage.
Using a small fixed sized highwatermark on a 256G RAM machine
turns out to call BBPtrim too often/unnecessary
The new decision is aligned with MAL-Admission, where we start
worrying after 80% of the memory is claimed by rss.
At that stage we try to reduce the minimum needed to get back on safe grounds.