During primary key check perform subgroupby only to groups with > 1 members.
Make sure that we are not doing extra work by performing subgroupby on groups
that have onlyone member. In practice this should mean only one subgroup on the hash values of a combined key, and a subgroup on the few collisions (if any).
This emulates a candidate list from groupby. If we ever support directly candidate lists in the groupby operator, this code can be simplified by avoiding the leftfetch joins.