setObject: fix BigDecimal handling after previous fix
This fixes Test_PStypes again.  Even if a double is rounded, this
doesn't mean it has all trailing zeros, since the representation simply
can't guarantee that.  So, just cut off excess digits like we did at the
previous implementation not to give the server any more than exactly the
specification of the decimal (it expects).