Implemented timeout callback function.
You can specify a read and write timeout on a stream (only implemented
(or even interesting) for socket streams) which causes the read/write
to get aborted after the specified timeout.  If no data has been
transferred, the function returns an error.  If you specify a callback
function, it is called when the timeout happens, and if the callback
function returns FALSE (0), the read/write is tried again.

The timeout value was changed from seconds to milliseconds.