add optional column list (s) in copy into statement
ie copy ... into table_name [ ( list with to be inserted columns ) ]
	from 'file or stdin' [ ( list of columns in the file/stdin ) ]
	using etc.

The second list should match the number of columns in the input file. If
some of the columns are missing in the first list, they will be skipped
during the parsing process.

Column types should match with the to be inserted column type.