Adding 5 SQLite sqllogic regression tests
The test files and the *.stable.out files are quite big, so it takes some time to execute:
Running in sql/test/SQLite_regress/sqllogictest
14:50:34> select1.test   14.084s OK    OK
14:50:50> select2.test   13.082s OK    OK
14:51:04> select3.test   46.272s OK    OK
14:51:54> select4.test   25.216s OK    OK
14:52:56> select5.test  236.130s OK    OK
No differences encountered during testing.

Stefan and Sjoerd agreed to check these in so they are shared and tested on all platforms.
If needed some of the tests (eg select5.test) could be tuned by reducing the number of SELECTs.