Use proper way to stop daemon.
When using "monetdbd stop $FARM", the running daemon is sent a
SIGTERM, after which the sending monetdbd waits a while until the
daemon has actually stopped.  When using start-stop-daemon, we can do
the same thing: send a SIGTERM, and wait.  start-stop-daemon does not
invoke an external program to stop the daemon as the old code
apparently assumed.

This fixes bug 6108.