Be consistent about rounding or truncating values.
There is now a symbol TRUNCATE_NUMBERS that, if defined, which it
currently is not, will cause the code to always truncate towards zero
when values need to be converted from higher precision (number of
decimals after the decimal point) to lower precision, including from
floating point to integer.  Without the symbol, we always round.
This checkin also fixes some rounding bugs (don't add 0.5 to a
negative number and then truncate toward zero).

This fixes bug 6193.

Also added test.