Use the complete has entries in the string heap.
Until now, the hash entries at the start of the string heap was
relative to the start of the last 64KiB page in the string heap.  From
now on, the hash entry is relative to the start of the heap.  This
means that we don't need to clear the hash table anymore when we cross
a 64KiB page boundary.  It is still the case that the first 64KiB page
is duplicate eliminated, so if the hash entry is within this first
page, we follow the linked list.  We now do that even if the string
heap is larger than one page.