Fix a bug in the reading of the properties file
For each line of the form

key_i=value_i

in the properties file we call the function setConfValForKey. If key_i is not in
list of properties, then setConfValForKey will return an error that the key is
not recognized and in that case we need to add the property to the list.