Correct heartbeat object and add source attribute
Heartbeats had a trailing ',' and the cpuload list was in quotes. The
first problem is an error, and heartbeats would not parse. The second
would not let the user interpret the cpuload attribute as a list of
numbers.

Previously the heartbeat objects had a "user" attribute that always had
the value "heartbeat". The trace objects did not have a "user"
attribute (i.e. printing it was commented). Now both kinds of object
have an attribute "source" with the values "heartbeat" and "trace"
respectively.